fixed pricing thoughout

+++ java/src/main/java/com/techelevator/Candy.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Candy extends Product{
 
-	public Candy(String name, double price, String slot) {
+	public Candy(String name, BigDecimal price, String slot) {
 		super(name, price, slot);
 	}
