+++ java/src/main/java/com/techelevator/Chips.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Chips extends Product{
 	
-	public Chips(String name, double price, String slot) {
+	public Chips(String name, BigDecimal price, String slot) {
 		super(name, price, slot);
 	}
