+++ java/src/main/java/com/techelevator/Drink.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Drink extends Product {
 
-	public Drink(String name, double price, String slot) {
+	public Drink(String name, BigDecimal price, String slot) {
 		super(name, price, slot);
 		}
