+++ java/src/main/java/com/techelevator/Gum.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Gum extends Product{
 
-	public Gum(String name, double price, String slot) {
+	public Gum(String name, BigDecimal price, String slot) {
 		super(name, price, slot);
 	}
